changes style from bootstrap to .scss and bem
