+++ package.json
@@ -9,8 +9,10 @@
     "axios": "^0.24.0",
     "bootstrap": "^5.1.3",
     "chart.js": "^3.6.0",
+    "install": "^0.13.0",
+    "node-sass": "^6.0.1",
+    "npm": "^8.1.3",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0",
     "react-chartjs-2": "^3.3.0",
     "react-dom": "^17.0.2",
     "react-icons": "^4.3.1",
