+++ src/components/AlertComponent/AlertComponent.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
+import './AlertComponent.scss'
 
 
 const AlertComponent = ({ title, variant }) => {
   return (
-    <div className={`alert alert-${variant} text-center`} role="alert">
-      {title}
+    <div class={`alert ${variant}-alert`}>
+      <h3 class="alert__text">{title}</h3>
     </div>
   )
 }
