+++ src/components/AlertComponent/AlertComponent.scss
@@ -1,0 +1,26 @@
+.alert {
+    width: 100%;
+    margin: 20px auto;
+    height: 30px;
+    position: relative;
+    border-radius: 5px;
+    box-shadow: 0 0 15px 5px #ccc;
+
+    .alert__text {
+        text-align: center;
+        font-size: 20px;
+    }
+}
+
+@mixin alert($name, $bgColor) {
+    $accentColor: darken($bgColor, 50);
+
+    .#{$name} {
+        background-color:#{$bgColor};
+        border-left: 5px solid $accentColor;
+
+    }
+}
+
+@include alert(success-alert, #a8f0c6);
+@include alert(danger-alert, #f7a7a3);
