--- src/components/Loader.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Loader = () => {
-    return (
-        <div className="loader">
-            <div className="spinner-grow" role="status"></div>
-        </div>
-    )
-}
-
-export default Loader
