+++ src/components/Loader/Loader.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import './Loader.scss'
+
+const Loader = () => {
+    return (
+        <div className="lds__default">
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+            <div></div>
+        </div>
+
+    )
+}
+
+export default Loader
