+++ src/components/index.js
@@ -1,6 +1,6 @@
-export { default as Header } from './Header';
-export { default as Paginate } from './Paginate';
-export { default as LineChart } from './LineChart';
-export { default as Loader} from './Loader'
-export { default as ModalWindow } from './ModalWindow'
-export { default as AlertComponent } from './AlertComponent'
+export { default as Header } from './Header/Header';
+export { default as Paginate } from './Paginate/Paginate';
+export { default as LineChart } from './LineChart/LineChart';
+export { default as Loader} from './Loader/Loader'
+export { default as ModalWindow } from './ModalWindow/ModalWindow'
+export { default as AlertComponent } from './AlertComponent/AlertComponent'
