--- src/index.css
@@ -1,53 +1,0 @@
-* {
-    padding: 0;
-    margin: 0;
-}
-
-
-.modal_window {
-    height: 100%;
-    width: 100%;
-    background-color: rgba(0, 0, 0, 0.4);
-    position: fixed;
-    top: 0;
-    left: 0;
-    transform: scale(0);
-    transition: all .5s ease;
-    z-index: 1000;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.modal_window.active {
-    transform: scale(1)
-}
-
-.loader,
-.paginate {
-    margin: 0 auto;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-table tr:hover {
-    cursor: pointer;
-    background-color: #2ecc71;
-}
-
-.box-shadow {
-    box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
-    transition: transform .5s ease;
-    color: white;
-    font-weight: bold;
-}
-
-.box-shadow:hover {
-    transform: scale(1.05);
-    cursor: pointer;
-}
-
-.link {
-    cursor: pointer;
-}
