+++ src/index.scss
@@ -1,0 +1,62 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@100&display=swap');
+
+
+* {
+    margin: 0px;
+    padding: 0px;
+    box-sizing: border-box;
+
+    body {
+        width: 100%;
+        min-height: 100vh;
+        font-family: 'Roboto', sans-serif;
+        font-size: 45px;
+    }
+}
+
+$bg-primary: #0d6efd;
+$bg-secondary: #198754;
+
+ul,
+li {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+.container {
+    width: 90%;
+    margin: 0 auto;
+}
+
+
+$bp : (mobile : 480px,
+    tablet : 768px,
+    desktop : 1440px,
+);
+
+@mixin query($display) {
+    @each $key, $value in $bp {
+        @if ($display==$key) {
+            @media (min-width: $value) {
+                @content;
+            }
+        }
+    }
+}
+
+@mixin button-mixins() {
+    border-radius: 3px;
+    padding: 5px 3px;
+    color: white;
+    background-color: $bg-secondary;
+    font-weight: bold;
+}
+
+@mixin flexible() {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+
