+++ src/pages/CryptoDetail/CryptoDetail.scss
@@ -1,0 +1,108 @@
+@import '../../index.scss';
+
+// here i added styles
+.crypto {
+    .crypto__info {
+        &--title {
+            text-align: center;
+        }
+
+        &--cards {
+            @include flexible();
+            gap: 4px;
+        }
+    }
+
+    .card {
+        @include flexible();
+        flex-direction: column;
+        gap: 10px;
+        overflow: hidden;
+        margin: 10px;
+        max-width: 130px;
+        height: 150px;
+        width: 100%;
+        background-color: $bg-primary;
+        color: #fff;
+        font-weight: bold;
+        font-size: 20px;
+        box-shadow: 0 0 5px rgba(0, 0, 0, 0.15);
+        border-radius: 5px;
+        padding: 5px;
+        cursor: pointer;
+        transition: all .5s ease;
+
+        &__line {
+            width: 100%;
+            height: 4px;
+            background: #fff;
+            border-radius: 2px;
+        }
+
+        &:hover {
+            transform: scale(1.05);
+        }
+
+    }
+
+    .card__description--link {
+        text-decoration: none;
+        color: white;
+    }
+
+}
+
+.crypto__button--container {
+    width: 400px;
+    margin: 7px auto;
+    @include flexible();
+    gap: 5px;
+
+    .crypto__button {
+        padding: 15px 20px;
+        background-color: $bg-primary;
+        border: none;
+        color: #fff;
+        font-size: 20px;
+        border-radius: 5px;
+        cursor: pointer;
+        transition: all .7s ease;
+
+        &:hover {
+            opacity: 0.5
+        }
+    }
+
+}
+
+.crypto__info--chart {
+    .chart__title {
+        text-align: center;
+
+    }
+}
+
+
+@include query(mobile) {
+    .crypto__info--container {
+        .crypto__info--cards {
+            display: none;
+        }
+    }
+
+    .crypto__info--title {
+        display: none;
+    }
+}
+
+@include query(tablet) {
+    .crypto__info--container {
+        .crypto__info--cards {
+            display: flex;
+        }
+    }
+
+    .crypto__info--title {
+        display: block;
+    }
+}
