+++ src/pages/Main/Main.scss
@@ -1,0 +1,94 @@
+@import '../../index.scss';
+
+.currency {
+    margin-top: 10px;
+}
+
+table {
+    border-collapse: collapse;
+    margin: 0 auto;
+    font-size: 14px;
+    font-family: sans-serif;
+    min-width: 100%;
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
+    transition: all .5s ease;
+
+
+    thead tr {
+        background-color: $bg-primary;
+        color: #ffffff;
+        text-align: left;
+    }
+
+    th,
+    td {
+        padding: 11px 15px;
+    }
+
+    tbody tr {
+        border-bottom: 1px solid #dddddd;
+        font-size: 20px;
+    }
+
+    tbody tr:nth-of-type(even) {
+        background-color: #f3f3f3;
+    }
+
+    tbody tr:last-of-type {
+        border-bottom: 2px solid $bg-primary;
+    }
+
+
+    tbody tr:hover {
+        font-weight: bold;
+        color: white;
+        background-color: $bg-primary;
+        cursor: pointer;
+    }
+
+    tbody tr:nth-of-type(even):hover {
+        font-weight: bold;
+        color: white;
+        background-color: $bg-primary;
+        cursor: pointer;
+    }
+
+    td i {
+        font-size: 20px;
+    }
+
+}
+
+//styles for children modal window
+.currency__info {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 50px;
+}
+
+.block__enter {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex-direction: column;
+    gap: 20px;
+
+    .enter__text {
+        width: 90%;
+        background: $bg-primary;
+        color: white;
+        height: 50px;
+        padding: 0;
+        font-weight: bold;
+        border-radius: 5px;
+        @include flexible();
+    }
+
+    .enter__input {
+        width: 90%;
+        height: 40px;
+        padding: 0;
+        border-radius: 5px;
+    }
+}
