+++ src/pages/index.js
@@ -1,2 +1,2 @@
-export { default as Main } from './Main';
-export { default as CryptoDetails } from './CryptoDetails';
+export { default as Main } from './Main/Main';
+export { default as CryptoDetails } from './CryptoDetail/CryptoDetails';
